added all seeder datas

+++ backend/db/migrations/20221019041028-create-spot.js
@@ -34,11 +34,11 @@ module.exports = {
         allowNull: false
       },
       latitude: {
-        type: Sequelize.DECIMAL,
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       longitude: {
-        type: Sequelize.DECIMAL,
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       name: {
@@ -49,7 +49,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       price: {
-        type: Sequelize.DECIMAL,
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       createdAt: {
